+++ scripts/git-subtree/keepass/add.ps1
@@ -1,0 +1,5 @@
+git subtree add `
+    -P workspace/plugins/keepass `
+    https://github.com/Zuoqiu-Yingyi/siyuan-plugin-keepass.git `
+    dev `
+    --squash