+++ scripts/git-subtree/keepass/pull.ps1
@@ -1,0 +1,4 @@
+git subtree pull `
+    -P workspace/plugins/keepass `
+    https://github.com/Zuoqiu-Yingyi/siyuan-plugin-keepass.git `
+    dev